pyproject.toml
[tool.poetry]
name = "datapipe-framework"
version = "2.0.0"
description = "Framework profesional para migración de datos Oracle a Multi-Cloud (GCP, AWS, Azure)"
authors = ["Tu Equipo <equipo@example.com>"]
readme = "README.md"
packages = [{include = "datapipe", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
pyyaml = "^6.0.1"
pydantic = "^2.5.3"
jinja2 = "^3.1.2"
oracledb = "^2.0.1"
fastavro = "^1.9.0"
google-cloud-bigquery = "^3.14.1"
google-cloud-storage = "^2.14.0"
google-cloud-secret-manager = "^2.18.0"
boto3 = "^1.34.0"
azure-storage-blob = "^12.19.0"
azure-identity = "^1.15.0"
structlog = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
isort = "^5.13.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
datapipe = "datapipe.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
